Factor out dwSetup to its own class
Also adjust the timestamp format to more ISO 8601
